+++ src/scratchpad/src/org/apache/poi/hemf/record/emf/HemfMisc.java
@@ -360,7 +360,7 @@ public class HemfMisc {
             colorRef = new HwmfColorRef();
             int size = colorRef.init(leis);
             brushHatch = HwmfHatchStyle.valueOf((int) leis.readUInt());
-            return size + 3 * LittleEndianConsts.INT_SIZE;
+            return size + 3L * LittleEndianConsts.INT_SIZE;
         }
 
         @Override
@@ -513,7 +513,7 @@ public class HemfMisc {
 
             int size = colorRef.init(leis);
 
-            return size + 4 * LittleEndianConsts.INT_SIZE;
+            return size + 4L * LittleEndianConsts.INT_SIZE;
         }
 
         @Override
@@ -792,7 +792,7 @@ public class HemfMisc {
             // An XForm object that defines a two-dimensional linear transform in logical units.
             // This transform is used according to the ModifyWorldTransformMode to define a new value for
             // the world-space to page-space transform in the playback device context.
-            int size = readXForm(leis, xForm);
+            long size = readXForm(leis, xForm);
 
             // A 32-bit unsigned integer that specifies how the transform specified in Xform is used.
             // This value MUST be in the ModifyWorldTransformMode enumeration
