+++ src/scratchpad/src/org/apache/poi/hslf/blip/WMF.java
@@ -43,7 +43,7 @@ public final class WMF extends Metafile {
             InputStream is = new ByteArrayInputStream( rawdata );
             Header header = new Header();
             header.read(rawdata, CHECKSUM_SIZE*getUIDInstanceCount());
-            long skipLen = header.getSize() + CHECKSUM_SIZE*getUIDInstanceCount();
+            long skipLen = header.getSize() + (long)CHECKSUM_SIZE*getUIDInstanceCount();
             long skipped = IOUtils.skipFully(is, skipLen);
             assert(skipped == skipLen);
